refactor: Move most of the http specific code in a separate module
